lara/col/climb: port Lara_HangTest to TRX
