+++ src/libtrx/game/lara/col/climb.c
@@ -1,3 +1,4 @@
+#include "config.h"
 #include "game/input.h"
 #include "game/lara.h"
 #include "game/lara/util.h"
@@ -6,6 +7,7 @@
 #define M_CLIMB_SHIFT 70
 
 #define M_LF_HANG 21
+#define M_LF_STOP_HANG 9
 #define M_LF_CLIMB_L_SHIFT_START 28
 #define M_LF_CLIMB_L_SHIFT_END 29
 #define M_LF_CLIMB_R_SHIFT 57
@@ -20,6 +22,7 @@ typedef enum {
 
 // TODO: move all dependent climb functions from misc.c here
 static bool M_GetClimbStatus(void);
+static void M_HangTest(ITEM *item, COLL_INFO *coll);
 static bool M_TestLadderRelease(ITEM *item);
 static M_CLIMB_RESULT M_TestClimbUpPos(
     const ITEM *item, int32_t front, int32_t right, int32_t *shift,
@@ -39,6 +42,148 @@ static bool M_GetClimbStatus(void)
 #else
     return Lara_GetLaraInfo()->climb_status;
 #endif
+}
+
+static void M_HangTest(ITEM *const item, COLL_INFO *const coll)
+{
+    coll->bad_pos = NO_BAD_POS;
+    coll->bad_neg = NO_BAD_NEG;
+    coll->bad_ceiling = 0;
+    Lara_GetCollisionInfo(item, coll);
+    const bool flag = coll->side_front.floor < 200;
+
+    item->gravity = false;
+    item->fall_speed = 0;
+    LARA_INFO *const lara = Lara_GetLaraInfo();
+    lara->move_angle = item->rot.y;
+
+    DIRECTION dir = Math_GetDirection(item->rot.y);
+    switch (dir) {
+    case DIR_NORTH:
+        item->pos.z += 2;
+        break;
+    case DIR_EAST:
+        item->pos.x += 2;
+        break;
+    case DIR_SOUTH:
+        item->pos.z -= 2;
+        break;
+    case DIR_WEST:
+        item->pos.x -= 2;
+        break;
+    default:
+        break;
+    }
+
+    coll->bad_pos = NO_BAD_POS;
+    coll->bad_neg = -STEPUP_HEIGHT;
+    coll->bad_ceiling = 0;
+    Lara_GetCollisionInfo(item, coll);
+
+#if TR_VERSION == 1
+    const bool enable_swing_cancel = g_Config.gameplay.enable_swing_cancel;
+#else
+    const bool enable_swing_cancel = true;
+    if (M_GetClimbStatus()) {
+        if (!g_Input.action || item->hit_points <= 0) {
+            XYZ_32 pos = {
+                .x = 0,
+                .y = 0,
+                .z = 0,
+            };
+            Collide_GetJointAbsPosition(item, &pos, 0);
+            if (dir == DIR_NORTH || dir == DIR_SOUTH) {
+                item->pos.x = pos.x;
+            } else {
+                item->pos.z = pos.z;
+            }
+
+            item->goal_anim_state = LS_JUMP_FORWARD;
+            item->current_anim_state = LS_JUMP_FORWARD;
+            Item_SwitchToAnim(item, LA_FALL_START, 0);
+            item->pos.y += STEP_L;
+            item->gravity = true;
+            item->speed = 2;
+            item->fall_speed = 1;
+            lara->gun_status = LGS_ARMLESS;
+            return;
+        }
+
+        if (!Lara_TestHangOnClimbWall(item, coll)) {
+            item->pos.x = coll->old.x;
+            item->pos.y = coll->old.y;
+            item->pos.z = coll->old.z;
+            item->goal_anim_state = LS_HANG;
+            item->current_anim_state = LS_HANG;
+            Item_SwitchToAnim(item, LA_REACH_TO_HANG, M_LF_HANG);
+            return;
+        }
+
+        if (Item_TestAnimEqual(item, LA_REACH_TO_HANG)
+            && Item_TestFrameEqual(item, M_LF_HANG)
+            && Lara_TestClimbStance(item, coll)) {
+            item->goal_anim_state = LS_CLIMB_STANCE;
+        }
+        return;
+    }
+#endif
+
+    if (!g_Input.action || item->hit_points <= 0
+        || coll->side_front.floor > 0) {
+        item->goal_anim_state = LS_JUMP_UP;
+        item->current_anim_state = LS_JUMP_UP;
+        Item_SwitchToAnim(item, LA_JUMP_UP, M_LF_STOP_HANG);
+        const BOUNDS_16 *const bounds = Item_GetBoundsAccurate(item);
+        // Retain original bug for #3134, whereby in TR2 Lara perpatually swing
+        // cancels on death.
+        if (enable_swing_cancel && (TR_VERSION >= 2 || item->hit_points > 0)) {
+            item->pos.y += bounds->max.y;
+        } else {
+            item->pos.y += coll->side_front.floor - bounds->min.y + 2;
+        }
+        item->pos.x += coll->shift.x;
+        item->pos.z += coll->shift.z;
+        item->gravity = true;
+        item->speed = 2;
+        item->fall_speed = 1;
+        lara->gun_status = LGS_ARMLESS;
+        return;
+    }
+
+    const BOUNDS_16 *const bounds = Item_GetBoundsAccurate(item);
+    const int32_t hdif = coll->side_front.floor - bounds->min.y;
+
+    if (ABS(coll->side_left.floor - coll->side_right.floor) >= SLOPE_DIF
+        || coll->side_mid.ceiling >= 0 || coll->coll_type != COLL_FRONT || flag
+        || coll->hit_static || hdif < -SLOPE_DIF || hdif > SLOPE_DIF) {
+        item->pos = coll->old;
+        if (item->current_anim_state == LS_SHIMMY_LEFT
+            || item->current_anim_state == LS_SHIMMY_RIGHT) {
+            item->goal_anim_state = LS_HANG;
+            item->current_anim_state = LS_HANG;
+            Item_SwitchToAnim(item, LA_REACH_TO_HANG, M_LF_HANG);
+        }
+        return;
+    }
+
+    switch (dir) {
+    case DIR_NORTH:
+    case DIR_SOUTH:
+        item->pos.z += coll->shift.z;
+        break;
+
+    case DIR_EAST:
+    case DIR_WEST:
+        item->pos.x += coll->shift.x;
+        break;
+
+    default:
+        break;
+    }
+
+    if (TR_VERSION >= 2 || (hdif >= -STEP_L && hdif <= STEP_L)) {
+        item->pos.y += hdif;
+    }
 }
 
 static bool M_TestLadderRelease(ITEM *const item)
@@ -164,7 +309,7 @@ static M_CLIMB_RESULT M_TestClimbUpPos(
 
 static void M_Hang(ITEM *const item, COLL_INFO *const coll)
 {
-    Lara_HangTest(item, coll);
+    M_HangTest(item, coll);
     if (item->goal_anim_state != LS_HANG) {
         return;
     }
@@ -206,7 +351,7 @@ static void M_Shimmy(ITEM *const item, COLL_INFO *const coll)
         item->current_anim_state == LS_SHIMMY_LEFT ? -DEG_90 : DEG_90;
     LARA_INFO *const lara = Lara_GetLaraInfo();
     lara->move_angle = item->rot.y + angle;
-    Lara_HangTest(item, coll);
+    M_HangTest(item, coll);
     lara->move_angle = item->rot.y + angle;
 }
