+++ src/tr1/game/lara/misc.c
@@ -15,104 +15,7 @@
 #include <stdint.h>
 
 #define LF_FASTFALL 1
-#define LF_STOPHANG 9
 #define LF_STARTHANG 12
-#define LF_HANG 21
-
-void Lara_HangTest(ITEM *item, COLL_INFO *coll)
-{
-    int flag = 0;
-    const BOUNDS_16 *bounds;
-
-    coll->bad_pos = NO_BAD_POS;
-    coll->bad_neg = NO_BAD_NEG;
-    coll->bad_ceiling = 0;
-    Lara_GetCollisionInfo(item, coll);
-    if (coll->side_front.floor < 200) {
-        flag = 1;
-    }
-
-    g_Lara.move_angle = item->rot.y;
-    item->gravity = false;
-    item->fall_speed = 0;
-
-    PHD_ANGLE angle = (uint16_t)(item->rot.y + DEG_45) / DEG_90;
-    switch (angle) {
-    case DIR_NORTH:
-        item->pos.z += 2;
-        break;
-
-    case DIR_WEST:
-        item->pos.x -= 2;
-        break;
-
-    case DIR_SOUTH:
-        item->pos.z -= 2;
-        break;
-
-    case DIR_EAST:
-        item->pos.x += 2;
-        break;
-    }
-
-    coll->bad_pos = NO_BAD_POS;
-    coll->bad_neg = -STEPUP_HEIGHT;
-    coll->bad_ceiling = 0;
-    Lara_GetCollisionInfo(item, coll);
-
-    if (!g_Input.action || item->hit_points <= 0) {
-        item->goal_anim_state = LS_JUMP_UP;
-        item->current_anim_state = LS_JUMP_UP;
-        Item_SwitchToAnim(item, LA_JUMP_UP, LF_STOPHANG);
-        bounds = Item_GetBoundsAccurate(item);
-        if (g_Config.gameplay.enable_swing_cancel && item->hit_points > 0) {
-            item->pos.y += bounds->max.y;
-        } else {
-            item->pos.y += coll->side_front.floor - bounds->min.y + 2;
-        }
-        item->pos.x += coll->shift.x;
-        item->pos.z += coll->shift.z;
-        item->gravity = true;
-        item->fall_speed = 1;
-        item->speed = 2;
-        g_Lara.gun_status = LGS_ARMLESS;
-        return;
-    }
-
-    bounds = Item_GetBoundsAccurate(item);
-    const int32_t hdif = coll->side_front.floor - bounds->min.y;
-
-    if (ABS(coll->side_left.floor - coll->side_right.floor) >= SLOPE_DIF
-        || coll->side_mid.ceiling >= 0 || coll->coll_type != COLL_FRONT
-        || hdif < -SLOPE_DIF || hdif > SLOPE_DIF || flag) {
-        item->pos.x = coll->old.x;
-        item->pos.y = coll->old.y;
-        item->pos.z = coll->old.z;
-        if (item->current_anim_state == LS_SHIMMY_LEFT
-            || item->current_anim_state == LS_SHIMMY_RIGHT) {
-            item->goal_anim_state = LS_HANG;
-            item->current_anim_state = LS_HANG;
-            Item_SwitchToAnim(item, LA_REACH_TO_HANG, LF_HANG);
-        }
-        return;
-    }
-
-    switch (angle) {
-    case DIR_NORTH:
-    case DIR_SOUTH:
-        item->pos.z += coll->shift.z;
-        break;
-
-    case DIR_WEST:
-    case DIR_EAST:
-        item->pos.x += coll->shift.x;
-        break;
-    }
-
-    if (hdif >= -STEP_L && hdif <= STEP_L) {
-        item->pos.y += hdif;
-    }
-}
 
 void Lara_SlideSlope(ITEM *item, COLL_INFO *coll)
 {
